tests: split kfp-api's integration tests into database and obs integration
To avoid mixing scenarios and conditions that may affect the tests,
specially when asserting for a blocked status after configuring the
charm in a certain way, this commit splits the integration test file
into two logical test suites that group related test cases: one for
observability and another one for databases.
